Rename google-pubsub Filebeat input to gcp-pubsub (#22213)
* Rename google-pubsub Filebeat input to gcp-pubsub

* change to gcp-pubsub input name in elastic-agent

* add warning log message for deprecated old input name

(cherry picked from commit 42428a8b20d8bcda9979884e89dea25890e02cb8)
