+++ x-pack/elastic-agent/spec/filebeat.yml
@@ -64,7 +64,7 @@ rules:
     - cloudfoundry
     - container
     - docker
-    - google-pubsub
+    - gcp-pubsub
     - http_endpoint
     - httpjson
     - kafka
