+++ x-pack/filebeat/_meta/config/filebeat.inputs.reference.xpack.yml.tmpl
@@ -30,7 +30,7 @@
 
 #---------------------------- Google Cloud Pub/Sub Input -----------------------
 # Input for reading messages from a Google Cloud Pub/Sub topic subscription.
-- type: google-pubsub
+- type: gcp-pubsub
   enabled: false
 
   # Google Cloud project ID. Required.
