+++ x-pack/filebeat/docs/inputs/input-gcp-pubsub.asciidoc
@@ -1,15 +1,15 @@
 [role="xpack"]
 
-:type: google-pubsub
+:type: gcp-pubsub
 
 [id="{beatname_lc}-input-{type}"]
-=== Google Cloud Pub/Sub input
+=== GCP Pub/Sub input
 
 ++++
-<titleabbrev>Google Pub/Sub</titleabbrev>
+<titleabbrev>GCP Pub/Sub</titleabbrev>
 ++++
 
-Use the `google-pubsub` input to read messages from a Google Cloud Pub/Sub topic
+Use the `gcp-pubsub` input to read messages from a Google Cloud Pub/Sub topic
 subscription.
 
 This input can, for example, be used to receive Stackdriver logs that have been
@@ -23,7 +23,7 @@ Example configuration:
 ["source","yaml",subs="attributes"]
 ----
 {beatname_lc}.inputs:
-- type: google-pubsub
+- type: gcp-pubsub
   project_id: my-gcp-project-id
   topic: vpc-firewall-logs-topic
   subscription.name: filebeat-vpc-firewall-logs-sub
@@ -33,7 +33,7 @@ Example configuration:
 
 ==== Configuration options
 
-The `google-pubsub` input supports the following configuration options plus the
+The `gcp-pubsub` input supports the following configuration options plus the
 <<{beatname_lc}-input-{type}-common-options>> described later.
 
 [float]
