+++ x-pack/filebeat/input/gcppubsub/_meta/Dockerfile
